chore(booking): move booking from legacy to pages
