--- src/legacy/routes/Booking/index.ts
@@ -1,1 +1,0 @@
-export { default } from './Booking';
