+++ src/pages/bookings/BookingPayment/BookingPayment.tsx
