+++ src/pages/bookings/BookingPayment/BookingPaymentBar.container.ts
