+++ src/pages/bookings/BookingPayment/BookingPaymentBar.tsx
