+++ src/pages/bookings/BookingPayment/index.ts
