+++ src/pages/bookings/BookingReceipt/BookingReceipt.container.ts
