+++ src/pages/bookings/BookingReceipt/BookingReceiptBar.container.ts
