+++ src/pages/bookings/BookingReceipt/BookingReceiptBar.tsx
