+++ src/pages/bookings/BookingReceipt/index.ts
